REFACTOR: Moved blog related functions into blog directory
